+++ CHANGELOG.md
@@ -11,9 +11,6 @@ breakage if you use `no-default-features`.
 There were/are numerous minor versions before 1.0 due to the language changes.
 Versions with only mechanical changes will be omitted from the following list.
 
-## 0.4.21 (unreleased)
-* Add `from_timestamp_millis`, `from_timestamp_micros`, `from_timestamp_nanos` for `NaiveDateTime`
-
 ## 0.4.20 (unreleased)
 
 * Add more formatting documentation and examples.
@@ -27,6 +24,7 @@ Versions with only mechanical changes will be omitted from the following list.
 * Add optional rkyv support.
 * Add support for microseconds timestamps serde serialization for `NaiveDateTime`.
 * Add support for optional timestamps serde serialization for `NaiveDateTime`.
+* Add `from_timestamp_millis`, `from_timestamp_micros`, `from_timestamp_nanos` for `NaiveDateTime` (#562)
 
 ## 0.4.19